В методе std::string Serialization(Node *node) после освобождения памяти сделал присвоение rootNode = nullptr;, чтобы нечаянно не обращаться к этой области памяти.

+++ src/NodeClass.hpp
@@ -50,7 +50,7 @@ enum DataType
 			return subNodes;
 		}
 
-		int getSubNodesCount() const
+		size_t getSubNodesCount() const
 		{
 			return subNodes.size();
 		}
